fix: updated the 'permissions' column type on managed table

closes #95

+++ src/accounts/entities/managed.entity.ts
@@ -43,7 +43,7 @@ export class Managed extends LinkedToAccount {
   })
   author?: Relation<Account>;
 
-  @Column('bigint', {
+  @Column('int', {
     unsigned: true,
     default: '0',
   })
